requirements.txt
scapy
mac-vendor-lookup
os
sys
time
netifaces
python-nmap
